@@ -14,33 +14,30 @@
   "dependencies": {
     "@angular-devkit/core": "~0.4.5",
     "@angular-devkit/schematics": "~0.4.5",
-    "@angular/animations": "~5.2.9",
-    "@angular/cdk": "~5.2.4",
-    "@angular/common": "~5.2.9",
-    "@angular/compiler": "~5.2.9",
-    "@angular/core": "~5.2.9",
-    "@angular/flex-layout": "^5.0.0-beta.13",
-    "@angular/forms": "~5.2.9",
-    "@angular/http": "~5.2.9",
-    "@angular/material": "~5.2.4",
-    "@angular/platform-browser": "~5.2.9",
-    "@angular/platform-browser-dynamic": "~5.2.9",
-    "@angular/router": "~5.2.9",
+    "@angular/animations": "~6.0.7",
+    "@angular/cdk": "~6.3.2",
+    "@angular/common": "~6.0.7",
+    "@angular/compiler": "~6.0.7",
+    "@angular/core": "~6.0.7",
+    "@angular/flex-layout": "^6.0.0-beta.16",
+    "@angular/forms": "~6.0.7",
+    "@angular/http": "~6.0.7",
+    "@angular/material": "~6.3.2",
+    "@angular/platform-browser": "~6.0.7",
+    "@angular/platform-browser-dynamic": "~6.0.7",
+    "@angular/router": "~6.0.7",
     "@schematics/angular": "~0.4.3",
     "@types/lodash": "^4.14.104",
     "@types/underscore": "^1.8.7",
-    "angularx-qrcode": "^1.1.7",
     "core-js": "^2.4.1",
     "fullcalendar": "3.6.1",
     "hammerjs": "^2.0.8",
     "jquery": "^3.3.1",
     "ng-fullcalendar": "^1.5.6",
-    "ng-pick-datetime": "5.2.6",
-    "ng-pick-datetime-moment": "1.0.5",
-    "print-js": "^1.0.43",
-    "qz-tray": "^2.0.6-1",
+    "npm": "^6.1.0",
     "react-scripts-ts": "^2.13.0",
-    "rxjs": "^5.5.6",
+    "rxjs": "^6.2.1",
+    "rxjs-compat": "^6.2.1",
     "set-manipulator": "^0.3.1",
     "sha": "^2.0.1",
     "underscore": "^1.8.3",
@@ -48,9 +45,10 @@
     "zone.js": "^0.8.20"
   },
   "devDependencies": {
-    "@angular/cli": "1.7.2",
-    "@angular/compiler-cli": "~5.2.9",
-    "@angular/language-service": "~5.2.9",
+    "@angular-devkit/build-angular": "~0.0.9",
+    "@angular/cli": "^6.0.8",
+    "@angular/compiler-cli": "~6.0.7",
+    "@angular/language-service": "~6.0.7",
     "@types/jasmine": "~2.8.6",
     "@types/jasminewd2": "~2.0.2",
     "@types/jquery": "^3.3.1",
@@ -67,6 +65,7 @@
     "protractor": "~5.1.2",
     "ts-node": "~3.2.0",
     "tslint": "~5.9.1",
-    "typescript": "~2.5.3"
+    "typescript": "~2.7.2",
+    "webpack": "^4.14.0"
   }
 }
