@@ -1,14 +1,13 @@
 import { Component, Input, OnInit, OnDestroy } from '@angular/core';
 import {ENTER, COMMA} from '@angular/cdk/keycodes';
 import { MatChipInputEvent, MatAutocompleteSelectedEvent } from '@angular/material';
-import { Observable } from 'rxjs/Observable';
+import { Observable, Subscription } from 'rxjs';
 
 import {  difference } from 'set-manipulator';
 
 import { Devotee, ServiceArea, DevoteeServiceInterest  } from '../../shared/sdk/models';
 import { ServiceAreaApi, DevoteeServiceInterestApi } from '../../shared/sdk';
 import { DevoteeSearchSelectService, NotificationService } from '../../shared/services';
-import { Subscription } from 'rxjs/Subscription';
 import {LoopBackFilter} from '../../shared/sdk/models/BaseModels';
 
 @Component({
