@@ -1,6 +1,5 @@
 import { Inject, Injectable } from '@angular/core';
-import { Observable } from 'rxjs/Observable';
-import 'rxjs/add/observable/of';
+import { Observable, of } from 'rxjs';
 
 import {LoopBackFilter} from '../../shared/sdk/models/BaseModels';
 import { AuthService } from '../../shared/services';
