@@ -1,6 +1,5 @@
 import { Injectable, OnInit, OnDestroy } from '@angular/core';
-import { Observable } from 'rxjs/Observable';
-import {BehaviorSubject} from 'rxjs/BehaviorSubject';
+import { Observable, BehaviorSubject } from 'rxjs';
 import {catchError, finalize} from 'rxjs/operators';
 
 
