@@ -8,11 +8,12 @@ import { LoopBackAuth } from './auth.service';
 import { LoopBackConfig } from '../../lb.config';
 import { LoopBackFilter, AccessToken } from '../../models/BaseModels';
 import { SDKModels } from '../custom/SDKModels';
-import { Observable } from 'rxjs/Observable';
+/* import { Observable } from 'rxjs/Observable';
 import { Subject } from 'rxjs/Subject';
-import { ErrorObservable } from 'rxjs/observable/ErrorObservable';
-import 'rxjs/add/operator/catch';
-import 'rxjs/add/operator/map';
+import { ErrorObservable } from 'rxjs/observable/ErrorObservable'; */
+import { Observable, Subject } from 'rxjs';
+import { map, catchError } from 'rxjs/operators';
+
 // Making Sure EventSource Type is available to avoid compilation issues.
 declare var EventSource: any;
 /**
