@@ -1,6 +1,5 @@
 import { Injectable } from '@angular/core';
-import { BehaviorSubject } from "rxjs/BehaviorSubject";
-import { Observable } from "rxjs/Observable";
+import { Observable, BehaviorSubject } from "rxjs";
 
 @Injectable()
 export class GlobalEventsManager {
