@@ -5,7 +5,7 @@ import {  NotificationService} from '../../shared/services';
 import { MaterialModule } from '../../material.module';
 import { MatPaginator, MatSort, MatTableDataSource, MatSelectChange } from '@angular/material';
 import {MatDialog} from '@angular/material';
-import { Subscription } from 'rxjs/Subscription';
+import { Subscription } from 'rxjs';
 
 import { SharedComponentsModule } from '../../shared/components/shared-components.module'
 import { InlineEditComponent } from '../../shared/components/inline-edit/inline-edit.component';
